refactor: Move routing related stuff to the routing folder
