+++ src/routing/Navigator.tsx
@@ -1,15 +1,15 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 import { Suspense } from 'react';
 
-import CenterLoadingSpinner from './components/CenterLoadingSpinner';
+import CenterLoadingSpinner from '../components/CenterLoadingSpinner';
 
 // These Pages are loaded initially:
-import Login from './pages/Login';
-import Welcome from './pages/Welcome';
-import LoginToken from './pages/LoginToken';
-import { RequireAuth } from './User';
-import FullPageModal from './modals/FullPageModal';
-import { lazyWithRetry } from './lazy';
+import Login from '../pages/Login';
+import Welcome from '../pages/Welcome';
+import LoginToken from '../pages/LoginToken';
+import { RequireAuth } from '../User';
+import FullPageModal from '../modals/FullPageModal';
+import { lazyWithRetry } from '../lazy';
 
 // All other pages load lazy:
 const NavigatorLazy = lazyWithRetry(() => import('./NavigatorLazy'), { prefetch: true });
